Move library model to common
