+++ .gitignore
@@ -3,3 +3,4 @@ venv/
 db.sqlite3
 __pycache__
 *.pyc
+*_initial.py
