+++ hantorism_library/views.py
@@ -1,8 +1,8 @@
 from django.shortcuts import render
-from .models import Library
+from common_hantorism.models import HantorismLibrary as Model
 
 
 def library(request):
-    books = Library.objects.all()
+    books = Model.objects.all()
     context = {'books': books}
     return render(request, 'library.html', context)
